+++ game/main.py
@@ -1,12 +1,10 @@
 from game.scrabble import ScrabbleGame
-from game.player import Player
-from game.game_models import BagTiles
 import random
 
 
 
 def main():
-        bagtiles = BagTiles()
+
         
         print("Bienvenidos al scrabble de luciano!")
         while True:
@@ -19,6 +17,7 @@ def main():
             except ValueError:
                 print("Error Ingrese de 1 a 4 jugadores")
         scrabbleGame = ScrabbleGame(players_count=players_count)
+        bag_tiles=scrabbleGame.bag_tiles
         
         print("Cantidad de jugadores: ", len(scrabbleGame.players))
 
@@ -29,41 +28,39 @@ def main():
 
         while True:
             player=scrabbleGame.players[scrabbleGame.current_player]
+            if len(player.rack)==0:
+                scrabbleGame.end_game()
+                
             print(f"Turno del jugador {scrabbleGame.players[scrabbleGame.current_player].nickname} \n")
+            player.display_rack()
             print('_' * 15)
             print('--Menu--')
-            print('1. Mostrar Tablero')
-            print('2. Mostrar puntos')
-            print('3. Mostrar mis fichas')
-            print('4. Colocar palabra')
-            print('5. Cambiar fichas')
-            print('6. Pasar turno')
-            print('7. Terminar juego')
-            print('+' * 100)
+            print('1. Tablero')
+            print('2. Puntos')
+            print('3. Colocar palabra')
+            print('4. Cambiar fichas')
+            print('5. Pasar turno')
+            print('6. Terminar juego')
             option = input("Ingrese una opción: ")
 
 
             if option == '1':
-                print ('tablero:')
                 scrabbleGame.show_board()
             elif option == '2':
                 scrabbleGame.show_score()
             elif option == '3':
-                player.display_rack()
-            elif option == '4':
                 scrabbleGame.play()
-            elif option == '5':
-                bagtiles.put(player.rack)
-                random.shuffle(bagtiles.bag)
-                player.rack = bagtiles.take(len(player.rack))
+            elif option == '4':
+                bag_tiles.put(player.rack)
+                random.shuffle(bag_tiles.bag)
+                player.rack = bag_tiles.take(len(player.rack))
                 rack_strings = [str(tile) for tile in player.rack]   
                 print("Sus nuevas fichas son : " + ", ".join(rack_strings))
-            elif option == '6':
+            elif option == '5':
                 scrabbleGame.next_turn()
-            elif option == '7':   
+            elif option == '6':   
                 scrabbleGame.end_game()
-
-
+            
             else:
                 print ('------------------------------------------------------------------------------------------------------\n'
             )